requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0
httpx==0.27.2
stripe==11.1.1
resend==2.4.0
boto3==1.35.36
python-dotenv==1.0.1
